pyproject.toml
[tool.poetry]
name = "pyicatu"
version = "0.1.0"
description = ""
authors = ["Antonio Bonna <antoniobonna@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
ruff = "^0.11.2"
dbt-core = "^1.9.3"
dbt-postgres = "^1.9.0"
yfinance = "^0.2.55"
pandas = "^2.2.3"
sqlalchemy = "^2.0.40"
python-dotenv = "^1.1.0"
requests = "^2.32.3"
python-dateutil = "^2.9.0.post0"
psycopg2-binary = "^2.9.10"
swifter = "^1.4.0"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
pydantic = "^2.11.1"

[build-system]
requires = ["poetry-core", "setuptools"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Enable ruff format
line-length = 100
target-version = "py312" # Replace with your target Python version (e.g., py39, py310)

[tool.ruff.lint]
# E501: line too long
# E/F: pycodestyle/Pyflakes errors
# W: pycodestyle warnings
# C: complexity (McCabe)
# I: isort
# N: naming
# D: pydocstyle
# UP: pyupgrade
# B: flake8-bugbear
# PL: Pylint
# RUF: Ruff-specific rules
select = [
    "E501",    # line length
    "Q",       # quotes
    "I",       # isort
    "PL",      # pylint
    "W291",    # trailing whitespace
    "W292",    # no newline at end of file
    "W293",    # blank line contains whitespace
]

[tool.ruff.format]
quote-style = "double"  # Optional: Use double quotes

[tool.ruff.lint.isort]
# Your project name
known-first-party = ["pyicatu"]

# Ruff automatically detects third-party packages from poetry dependencies
known-third-party = []

section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]
relative-imports-order = "closest-to-furthest"
combine-as-imports = true
split-on-trailing-comma = false